2021/src/test/java/Day5Test.java
import static org.junit.jupiter.api.Assertions.assertEquals;

import org.junit.jupiter.api.Nested;
import org.junit.jupiter.api.Test;

class Day5Test {

  @Nested
  class Teaser {

    Day puzzle = new Day5(
        "src/main/resources/day5-part1-teaser.txt");

    @Test
    void getAnswerPart1() {
      assertEquals(puzzle.getAnswerPart1(), 5);
    }

    @Test
    void getAnswerPart2() {
      assertEquals(puzzle.getAnswerPart2(), 12);
    }
  }

  @Nested
  class Real {

    Day puzzle = new Day5(
        "src/main/resources/day5.txt");

    @Test
    void getAnswerPart1() {
      assertEquals(puzzle.getAnswerPart1(), 6856);
    }

    @Test
    void getAnswerPart2() {
      assertEquals(puzzle.getAnswerPart2(), 20666);
    }
  }
}